Fix duration type and StrengthActivity issues
Change Activity::getDuration() to return double for proper precision; add override specifiers to getType()/describe() in CardioActivity and StrengthActivity headers; reorder StrengthActivity initializer list to match member declaration order; fix typo in output label ("[Strength]"); and correct types used for sets/reps in stream parsing (int instead of double). These changes address type/precision correctness, avoid initialization-order warnings, and fix I/O/formatting bugs.

+++ fittrack/activities/Activity.cpp
@@ -22,7 +22,7 @@ Activity::Activity(const std::string& date, double dur, double cal): date(date),
 Activity::~Activity(){}
 
 double Activity::getCalories() const{ return calories; }
-int Activity::getDuration() const{ return duration; }
+double Activity::getDuration() const{ return duration; }
 std::string Activity::getDate() const{ return date; }
 
 std::ostream& operator<<(std::ostream& out, const Activity& a){
